+++ demo/demo/views.py
@@ -12,19 +12,23 @@ def custom_admin_view(request):
     # request.current_app = 'admin'
     """
     context = admin.site.each_context(request)
-    context.update({
-        'title': 'Custom view',
-    })
+    context.update(
+        {
+            "title": "Custom view",
+        }
+    )
 
-    template = 'admin/custom_view.html'
+    template = "admin/custom_view.html"
     return render(request, template, context)
 
 
 def home(request):
-    return HttpResponse("""<html><head>
+    return HttpResponse(
+        """<html><head>
     <style>body {font-family: sans-serif; margin: 5rem; text-align: center; line-height: 1.7; font-size: 18px;}</style>
     </head>
     <body>
      Welcome to Django Suit demo app.<br>
      Go to <a href="/admin/">admin</a> to explore all the features.
-    </body></html>""")
+    </body></html>"""
+    )