+++ suit/__init__.py
@@ -1,2 +1,2 @@
-VERSION = '2.0beta3'
-default_app_config = 'suit.apps.DjangoSuitConfig'
+VERSION = "2022.05.03"
+default_app_config = "suit.apps.DjangoSuitConfig"