+++ suit/tests/test_menu.py
@@ -1,4 +1,5 @@
 from django.test import TestCase
+
 from suit.menu import ParentItem
 
 
@@ -10,4 +11,4 @@ class TestMenuTestCase(TestCase):
         """
         Just a dummy test case to satisfy Travis until all the tests are finished
         """
-        self.assertEqual(ParentItem(app='abc')._key(), 'abc')
+        self.assertEqual(ParentItem(app="abc")._key(), "abc")